% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_ovlpts.R
\name{buffer_ovlpts}
\alias{buffer_ovlpts}
\title{Move points that are too close}
\usage{
buffer_ovlpts(spacing, pt.buffer = NULL, bounds, pt.coords)
}
\arguments{
\item{spacing}{numeric of length 1, specifying minimum distance
(in units of bounds/pt.coords) that points should be apart.}

\item{pt.buffer}{numeric of length 1, specifying minimum distance
(in units of bounds/pt.coords) that points should be away from true
locations}

\item{bounds}{either object of class extent, or a numeric vector of
length 4, specifying the ymin, ymax, xmin and xmax bounds of the
points.}

\item{pt.coords}{Coordinates of points (x,y)}

\item{...}{Not currently in use}
}
\value{
A four column data.table specifying the
original and buffered point locations
}
\description{
\code{buffer_ovlpts} Buffer overlapping points along a
grid.
}
\details{
...
}
\examples{
\dontrun{
none yet
}
}
