% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geodata}
\alias{geodata}
\title{geodata}
\format{A set of 11 objects in unprojected lat/lon
coordinates. All datasets are either raster (bio.layers)
or SpatialLines/PolygonDataframes and have been cropped
to the extent of the diversity points.
\describe{
  \item{bio.layers}{RasterBrick with the 19 bioclim laaaters at
  a resolution of 2.5.}
  \item{extent}{The latitude / longitude extent of the dataset}
  \item{l.states}{SpatialPolygonsDataFrame with bounds of
  USA, MEX and CAN}
  \item{o.states}{SpatialPolygonsDataFrame with bounds of
  all non-USA admin}
  \item{us.states}{SpatialPolygonsDataFrame with bounds of
  USA states}
  \item{rivers}{SpatialLinesDataFrame with paths of rivers}
  \item{countries}{SpatialLinesDataFrame with boundaries of
  countries}
  \item{coasts}{SpatialLinesDataFrame with paths of coasts}
  \item{great.lakes}{SpatialPolygonsDataFrame with bounds of
  great lakes}
  \item{lakes}{SpatialPolygonsDataFrame with bounds of lakes}
  \item{land}{SpatialPolygonsDataFrame with bounds of land}
}}
\source{
\dontrun{
# 1. get and crop bioclim data
extent <- c(
  -111, -65, # longitude
  22, 51) # latitude

bio.layers <- raster::getData(
  name = 'worldclim',
  var = 'bio',
  res = 2.5)


# 2. get and crop other geographical data
devtools::install_github("jtlovell/spatialMapData")
library(spatialMapData)

# 2.1 states
data(
  ne_10m_admin_1_states_provinces.shp,
  package = "spatialMapData")
states <- crop(
  ne_10m_admin_1_states_provinces.shp,
  extent)
us.states1 <- subset(
  states,
  grepl("USA", adm1_code))
o.states <- subset(
  states,
  !grepl("USA", adm1_code))
l.states <- subset(
  states,
  grepl("MEX|USA|CAN", adm1_code))

# 2.2 rivers
data(
  ne_10m_rivers_lake_centerlines.shp,
  package = "spatialMapData")
rivers <- crop(
  ne_10m_rivers_lake_centerlines.shp,
  extent)

# 2.3 countries
data(
  ne_10m_admin_0_boundary_lines_land.shp,
  package = "spatialMapData")
countries <- crop(
  ne_10m_admin_0_boundary_lines_land.shp,
  extent)

# 2.4 coasts
data(
  ne_10m_coastline.shp,
  package = "spatialMapData")
coasts <- crop(
  ne_10m_coastline.shp,
  extent)

# 2.5 lakes
data(
  ne_10m_lakes.shp,
  package = "spatialMapData")
lakes <- crop(
  ne_10m_lakes.shp,
  extent)
great.lakes <- subset(
  lakes,
  lakes[[5]] == "Great Lakes")

# 2.6 land
data(
  ne_10m_land.shp,
  package = "spatialMapData")
land <- crop(
 ne_10m_land.shp,
  extent)

geodata <- list(
land = land,
lakes = lakes,
great.lakes = great.lakes,
coasts = coasts,
countries = countries,
rivers = rivers,
states = states,
o.states = o.states,
l.states = l.states,
extent = extent,
bio.layers = bio.layers)
save(
  geodata,
  file = "./data/geodata.rda")
 #' }
}
\usage{
geodata
}
\description{
A multi-object dataset contain many of the layers you
may want to plot for switchgrass diversity
}
\keyword{datasets}
