% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_layers.R
\name{project_layers}
\alias{project_layers}
\title{Project geodata layers}
\usage{
project_layers(proj = NULL, extent = c(-108.53, -64.51618, 20.76168,
  52.5362), geodata, crop.raster = TRUE, layers = "all")
}
\arguments{
\item{proj}{Projection to use. Otherwise defaults to WGS84,
centered on the middle of the extent}

\item{extent}{a numeric vector of length 4, specifying
the ymin, ymax, xmin and xmax bounds of the map}

\item{geodata}{named list containing geographic data layers. For a
layer to be projected and returned, it must have a class of either
Raster or Spatial.}

\item{crop.raster}{logical, should rasters be cropped? Speeds
things up a lot.}

\item{layers}{vector of geodata names to use. Defaults to all.}

\item{...}{Not currently in use}
}
\value{
A four column data.table specifying the
original and buffered point locations
}
\description{
\code{project_layers} Project layers for plotting
}
\details{
...
}
\examples{
\dontrun{
data("geodata", package = "pvdiv.map")
proj.layers <- project_layers(geodata = geodata)
projection <- proj.layers$projection
proj.layers <- proj.layers$proj.layers
plot(proj.layers[[1]])
}
}
